+++ structs.h
@@ -16,20 +16,20 @@ struct Title
 struct Node
 {
     std::string id;
-    double x_coord;
-    double y_coord;
+    long double x_coord;
+    long double y_coord;
     char *description;
     char *tag;
-    double elevation;
+    long double elevation;
     NodeType type;
-    double demand;
+    long double demand;
     std::string pattern;
-    double head;
-    double init_level;
-    double min_level;
-    double max_level;
-    double diameter;
-    double min_vol;
+    long double head;
+    long double init_level;
+    long double min_level;
+    long double max_level;
+    long double diameter;
+    long double min_vol;
     std::string vol_curve;
     std::string overflow;
 };
@@ -48,7 +48,7 @@ struct Annotation
 struct Pattern
 {
     std::string id;
-    std::vector<double> multipliers;
+    std::vector<long double> multipliers;
 };
 
 struct Pipe
@@ -59,17 +59,17 @@ struct Pipe
     int id;
     LineType lineType;
 
-    double length;
-    double diameter;
-    double roughness;
-    double minor_loss;
+    long double length;
+    long double diameter;
+    long double roughness;
+    long double minor_loss;
     PipeStatus pipe_status;
-    Pipe(Node *start, Node *end, const int &id, const LineType &lineType, const double &length, const double &diameter, const double &roughness, const double &minor_loss, const PipeStatus &pipe_status)
+    Pipe(Node *start, Node *end, const int &id, const LineType &lineType, const long double &length, const long double &diameter, const long double &roughness, const long double &minor_loss, const PipeStatus &pipe_status)
         : start(start), end(end), id(id), lineType(lineType), length(length), diameter(diameter), roughness(roughness), minor_loss(minor_loss), pipe_status(pipe_status)
     {
     }
 
-    double getAngle() const
+    long double getAngle() const
     {
         return atan2(end->y_coord - start->y_coord, end->x_coord - start->x_coord);
     }
@@ -93,9 +93,9 @@ struct Valve
     int id;
     Node *start;
     Node *end;
-    double diameter;
+    long double diameter;
     ValveType type;
-    double outlet_pressure;
+    long double outlet_pressure;
 };
 
 struct Tag
@@ -105,14 +105,14 @@ struct Tag
 struct Demand
 {
     Node *junction;
-    double demand;
+    long double demand;
     Pattern *pattern;
     std::string category;
 };
 
 struct Emitter {
     Node *junction;
-    double coefficient;
+    long double coefficient;
 };
 
 struct Status
@@ -125,17 +125,17 @@ struct Options
 {
     FlowUnit flow_unit = FlowUnit::UNDEFINED;
     HeadlossModel headloss_model = HeadlossModel::UNDEFINED;
-    double specific_gravity = 1.0;
-    double specific_viscosity = 1.0;
+    long double specific_gravity = 1.0;
+    long double specific_viscosity = 1.0;
     int maximum_trials = 1;
-    double accuracy = 0.001;
+    long double accuracy = 0.001;
 };
 
 struct Curve
 {
     int id;
-    double x_value;
-    double y_value;
+    long double x_value;
+    long double y_value;
 };
 
 struct Control
@@ -145,8 +145,8 @@ struct Control
 struct Leakage
 {
     int link_id;
-    double coefficient_1;
-    double coefficient_2;
+    long double coefficient_1;
+    long double coefficient_2;
 };
 
 struct Rule
@@ -154,15 +154,15 @@ struct Rule
 };
 struct Energy
 {
-    double global_efficiency;
-    double global_price;
-    double demand_charge;
+    long double global_efficiency;
+    long double global_price;
+    long double demand_charge;
 };
 
 struct Quality
 {
     Node *node;
-    double initial_quality;
+    long double initial_quality;
 };
 
 struct Source
@@ -177,7 +177,7 @@ struct Reaction
 {
     std::string type;
     std::string pipe_or_tank;
-    double coefficient;
+    long double coefficient;
 };
 
 struct Reaction_2
